requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.8
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
flatbuffers==25.2.10
fonttools==4.59.0
fqdn==1.5.1
fsspec==2025.7.0
gast==0.6.0
gevent==25.5.1
geventhttpclient==2.2.1
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==7.4.9
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==2.15.0
kiwisolver==1.4.8
lark==1.2.2
libclang==18.1.1
locust==2.27.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml-dtypes~=0.2.0
msgpack==1.1.1
namex==0.1.0
narwhals==1.48.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.4
notebook_shim==0.2.4
numpy==1.26.4
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.1
overrides==7.7.0
packaging==24.2
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rich-toolkit==0.14.8
rpds-py==0.26.0
scikit-learn==1.4.2
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.37.2
streamlit==1.35.0
tenacity==8.5.0
tensorboard~=2.15.0
tensorboard-data-server==0.7.2
tensorflow-macos==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.30.1
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt
zope.event==5.1.1
zope.interface==7.2
